Fix `username` field's vulnerability to a directory traversal attack.
The server's `username` field is vulnerable to directory traversal. This
means that a malicious user can enter a username such as `../test` and
instead of creating a `test.json` file in the directory where the server
is running, the file will actually be created in the directory *above*
where the server is running, because the `../` in the username will be
treated like a path, rather than part of the username. This can be
chained, allowing a malicious user to write files *anywhere* on the
server's filesystem that the operating system user that is running the
server can write to.

For more information about directory traversal attacks, sometimes called
path traversal attacks, see:

    https://www.owasp.org/index.php/Path_Traversal

This commit fixes the issue by processing the submitted username value
with Python's `os.path.basename()` method, which strips all but the
final element of a path string.