Prevent unauthenticated users from deleting other user's profiles.
This fixes a missing function-level access control vulnerabilty that
would allow a malicious user to delete the tests, profile information,
and reset the passwords of existing users by creating a new user with
the same name as the victim user. The vulnerability existed because the
server failed to check for the existence of the file that stored a given
user's profile information before creating a new, empty file with the
same name when asked.

This vulnerability also allows a malicious user to hijack the persona of
another user in the system because one of the fields that are
overwritten in the storage file is the victim user's password.

For more information about missing function-level access controls, see:
https://www.owasp.org/index.php/Top_10_2013-A7-Missing_Function_Level_Access_Control